# apps/web/static/generated/openapi/openfoundry.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenFoundry Proto-Derived API",
    "version": "0.1.0",
    "description": "Auto-generated JSON/HTTP contract derived from OpenFoundry proto services."
  },
  "paths": {
...
    "/api/v2/admin/control-panel": {
      "get": {
        "summary": "Get control panel (v2)",
        "operationId": "rest.admin.v2.getControlPanel",
        "tags": [
          "rest.admin.v2"
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ControlPanelSettings"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update control panel (v2)",
        "operationId": "rest.admin.v2.updateControlPanel",
        "tags": [
          "rest.admin.v2"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateControlPanelRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ControlPanelSettings"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/admin/groups": {
      "get": {
        "summary": "List groups (v2)",
        "operationId": "rest.admin.v2.listGroups",
        "tags": [
          "rest.admin.v2"
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminGroupsListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create group (v2)",
        "operationId": "rest.admin.v2.createGroup",
        "tags": [
          "rest.admin.v2"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGroupRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/admin/permissions": {
      "get": {
        "summary": "List permissions (v2)",
        "operationId": "rest.admin.v2.listPermissions",
        "tags": [
          "rest.admin.v2"
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminPermissionsListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create permission (v2)",
        "operationId": "rest.admin.v2.createPermission",
        "tags": [
          "rest.admin.v2"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePermissionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Permission"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/admin/policies": {
      "get": {
        "summary": "List policies (v2)",
        "operationId": "rest.admin.v2.listPolicies",
        "tags": [
          "rest.admin.v2"
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminPoliciesListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create policy (v2)",
        "operationId": "rest.admin.v2.createPolicy",
        "tags": [
          "rest.admin.v2"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertPolicyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Policy"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/admin/roles": {
      "get": {
        "summary": "List roles (v2)",
        "operationId": "rest.admin.v2.listRoles",
        "tags": [
          "rest.admin.v2"
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminRolesListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create role (v2)",
        "operationId": "rest.admin.v2.createRole",
        "tags": [
          "rest.admin.v2"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoleRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AdminGroupsListResponse": {
        "type": "object",
        "properties": {
          "count": {
            "type": "integer",
            "format": "int64"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GroupResponse"
            }
          }
        }
      },
      "AdminPermissionsListResponse": {
        "type": "object",
        "properties": {
          "count": {
            "type": "integer",
            "format": "int64"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Permission"
            }
          }
        }
      },
      "AdminPoliciesListResponse": {
        "type": "object",
        "properties": {
          "count": {
            "type": "integer",
            "format": "int64"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Policy"
            }
          }
        }
      },
      "AdminRolesListResponse": {
        "type": "object",
        "properties": {
          "count": {
            "type": "integer",
            "format": "int64"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoleResponse"
            }
          }
        }
      },
...
      "AppBrandingSettings": {
        "type": "object",
        "properties": {
          "accent_color": {
            "type": "string"
          },
          "display_name": {
            "type": "string"
          },
          "favicon_url": {
            "type": "string"
          },
          "logo_url": {
            "type": "string"
          },
          "primary_color": {
            "type": "string"
          },
          "show_powered_by": {
            "type": "boolean"
          }
        }
      },
...
      "ControlPanelSettings": {
        "type": "object",
        "properties": {
          "allow_self_signup": {
            "type": "boolean"
          },
          "allowed_email_domains": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "announcement_banner": {
            "type": "string"
          },
          "default_app_branding": {
            "$ref": "#/components/schemas/AppBrandingSettings"
          },
          "default_region": {
            "type": "string"
          },
          "deployment_mode": {
            "type": "string"
          },
          "docs_url": {
            "type": "string"
          },
          "maintenance_mode": {
            "type": "boolean"
          },
          "platform_name": {
            "type": "string"
          },
          "release_channel": {
            "type": "string"
          },
          "restricted_operations": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status_page_url": {
            "type": "string"
          },
          "support_email": {
            "type": "string"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_by": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "CreateGroupRequest": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "role_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        }
      },
...
      "CreatePermissionRequest": {
        "type": "object",
        "properties": {
          "action": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "resource": {
            "type": "string"
          }
        }
      },
...
      "CreateRoleRequest": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "permission_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        }
      },
...
      "GroupResponse": {
        "type": "object",
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "description": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "member_count": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "role_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "Permission": {
        "type": "object",
        "properties": {
          "action": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "description": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "resource": {
            "type": "string"
          }
        }
      },
...
      "Policy": {
        "type": "object",
        "properties": {
          "action": {
            "type": "string"
          },
          "conditions": {
            "type": "object",
            "additionalProperties": {}
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "created_by": {
            "type": "string",
            "format": "uuid"
          },
          "description": {
            "type": "string"
          },
          "effect": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "resource": {
            "type": "string"
          },
          "row_filter": {
            "type": "string"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "RoleResponse": {
        "type": "object",
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "description": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "permission_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "UpdateControlPanelRequest": {
        "type": "object",
        "properties": {
          "allow_self_signup": {
            "type": "boolean"
          },
          "allowed_email_domains": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "announcement_banner": {
            "type": "string"
          },
          "default_app_branding": {
            "$ref": "#/components/schemas/AppBrandingSettings"
          },
          "default_region": {
            "type": "string"
          },
          "deployment_mode": {
            "type": "string"
          },
          "docs_url": {
            "type": "string"
          },
          "maintenance_mode": {
            "type": "boolean"
          },
          "platform_name": {
            "type": "string"
          },
          "release_channel": {
            "type": "string"
          },
          "restricted_operations": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status_page_url": {
            "type": "string"
          },
          "support_email": {
            "type": "string"
          }
        }
      },
...
      "UpsertPolicyRequest": {
        "type": "object",
        "properties": {
          "action": {
            "type": "string"
          },
          "conditions": {
            "type": "object",
            "additionalProperties": {}
          },
          "description": {
            "type": "string"
          },
          "effect": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "resource": {
            "type": "string"
          },
          "row_filter": {
            "type": "string"
          }
        }
      },
...
    }
  }
}